Introduce Order Creation Functionality
In order to allow users to create orders via the Intelligent Foods API,
we need to implement the required functionality of build and submitting
order creation requests into the library. This PR introduces a few
classes that are used to build and process orders, including the order
item and the recipient.

This change addresses the need by:
* Introducing IntelligentFoods::OrderItem
* Introducing IntelligentFoods::Recipient