Update Partner API Endpoints
The endpoints for the authentication and api have recently been
changed. We should update the gem to use the new endpoints,
simplifying the configuration logic. In addition, the method in which
the Authentication is performed has changed, so we should also update
our logic to reflect those changes.

This change addresses the need by:
* Updating the endpoints
* Updating how the ApiClient authenticates