Added RSpec Testing Configuration & Fake
In order to allow customers to test against the gem, we should provide
them with a fake. This PR introduces a Testing module which configures
Rspec and substitutes a fake in place of the IntelligentFoods gem. Users
simply need to include 'inteligent_foods/rspec' in their spec helper.

This change addresses the need by:
* Introducing a rspec configuration
* Introducing a Testing module, and a fake