Replace Default Callback  with Shipment Created
The Intelligent Foods Partner API's callback response has changed, and
we should reflect those changes in our testing responses. Currently, the
default response is invalid and should be updated with the much more
useful, Shipment Created response. This will prove useful for users who
are testing the webhook feature of the partner api.

This change addresses the need by:
* Replacing the default callback with the shipment created callback
  response body